Make metadata XML required for EDI data publication

This is a wrapper function and utilizes the package EMLassemblyline to facilitate the creation of an EML document

make_eml_edi(
  data_file_names,
  attributes_file_names,
  title,
  table_names,
  table_descriptions,
  maintenance,
  edi_number
)

Arguments

data_file_names

A list of file names for data tables to be included in the publication. For instance, "catch.csv" or c("catch.csv", "trap.csv")

attributes_file_names

A list of file names for the attributes metadata csv templates. The naming convention for the files should be attributes_NAME_OF_DATA_TABLE.csv

title

Title of the data publication. The title should be clear, concise, and descriptive.

table_names

Short descriptive name for the data tables. If there are more than one data tables, provide as a vector of character strings. This needs to be in the same order as the data_file_names. This is an optional.

maintenance

This describes how often the data publication will be updated. The options are: daily, weekly, monthly, annually, complete.

edi_number

The EDI number to be used for the data publication. If the EML is generated to update an existing package, the number for the next version should be the input. For instance, if the existing EDI package is edi.1234.1 then the you should input edi.1234.2 because you are creating the EML for the next version. If this is for a new package, please reserve a new EDI number. Refer to the Quickstart Guide.

table_description

Description of the data table to help users understand what each data table contains. This needs to be in the same order as the data_file_names. This is an optional.

Value

An EML file within data-raw/eml. EMLassemblyline will include messages to ensure the EML file validates and will also flag any issues.

Details

This function assumes that all appropriate metadata templates have been completed Check that the following files in publish/metadata have been manually filled in and are correct and generate any additional files as needed: abstract.txt attributes/attributes_csv_template.csv (one per data table) custom_units.txt geographic_coverage.txt intellectual_rights.txt (DO NOT CHANGE unless using another license) keywords.txt methods.docx personnel/personnel_csv_template.csv

This function also assumes that the user has an EDI account and credentials are stored in their .Renviron as EDI_USER_ID

Examples

if (FALSE) { # \dontrun{
make_eml_edi(data_file_names = c("microhabitat_observations.csv","survey_locations.csv"),
             attributes_file_names = c("attributes_microhabitat_observations.csv","attributes_survey_locations.csv"),
             title = "Distribution and habitat use of juvenile steelhead and other fishes of the lower Feather River",
             table_names = c("Microhabitat Observations", "Survey Location Metadata"),
             table_descriptions = c("Primary data table with observations of microhabitat characteristics", "Description of the locations surveyed"),
             maintenance = "complete",
             edi_number = "edi.test")
} # }